requirements.txt
boto3==1.34.34
requests==2.31.0
beautifulsoup4==4.12.3
scrapy==2.11.0
mpi4py==3.1.5
python-dotenv==1.0.1
whoosh==2.7.4
celery==5.3.6
redis==5.0.1
urllib3==2.2.1
